회원 USER, CEO 구분하기 구현중,,

+++ src/main/java/sijang/sijanggaza/repository/ItemRepository.java
@@ -3,5 +3,6 @@ package sijang.sijanggaza.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sijang.sijanggaza.domain.Item;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
+
 }
